Reorganizacion de estructura del proyecto al mover el Footer a la carpeta de pages, Identacion de codigo y borrado de comentarios incesarios para una mejor lectura del codigo por Fernando Romero, Natali Suarez, Juan Baca, Juan Carrizo y Gabriel Calisaya
