+++ src/components/ProteccionDeRutas.jsx
@@ -1,9 +1,6 @@
 import { useSelector } from 'react-redux';
 import { Navigate, useLocation } from 'react-router-dom';
 
-/**
- * Este es el componente de Ruta Protegida. Su función es asegurarse de que solo los usuarios autenticados puedan acceder a ciertas rutas. Si el usuario no está autenticado, será redirigido a la página de inicio de sesión
- */
 
 const ProtectedRoute =({ children }) => {
     //Obtener el estado de autenticación del usuario desde Redux
@@ -11,7 +8,7 @@ const ProtectedRoute =({ children }) => {
     //Obtener la ubicación actual para guardar la ruta a la que se intentó acceder
     const location = useLocation();
 
-    //Si el usuario NO está autenticado, redirige a la página de inicio de sesión
+    //Si el usuario NO esta autenticado, redirige a la página de inicio de sesión
     //Se usa 'state: { from: location }' para guardar la ruta original y poder redirigir al usuario de vuelta a esa página después de iniciar sesión.
     
     if(!isAuthenticated) {
