+++ src/data/participantes.json
@@ -1,4 +1,5 @@
-[{
+[
+  {
     "nombre": "Calisaya Gabriel",
     "github": "GabrielCalisaya",
     "url": "https://github.com/GabrielCalisaya",
