+++ src/data/usersData.json
@@ -1,45 +1,47 @@
-[{
-    "nombre":"Gustavo Sosa",
-    "usuario":"gustavososa",
-    "password":"1234",
-    "role":"ADMIN",
+[
+  {
+    "nombre": "Gustavo Sosa",
+    "usuario": "gustavososa",
+    "password": "1234",
+    "role": "ADMIN",
     "isAuthenticated": true,
     "loginError": false,
     "img": "https://github.com/gustso.png"
-},
-{
-    "nombre":"Fernando Romero",
-    "usuario":"android",
-    "password":"4321",
-    "role":"USER",
+  },
+  {
+    "nombre": "Fernando Romero",
+    "usuario": "android",
+    "password": "4321",
+    "role": "USER",
     "isAuthenticated": true,
     "loginError": false,
     "img": "https://github.com/TuAndroide.png"
-},
-{
-    "nombre":"Juan Sebastian",
-    "usuario":"juanedm",
-    "password":"5678",
-    "role":"ADMIN",
+  },
+  {
+    "nombre": "Juan Sebastian",
+    "usuario": "juanedm",
+    "password": "5678",
+    "role": "ADMIN",
     "isAuthenticated": true,
     "loginError": false,
     "img": "https://github.com/Juansebaca.png"
-},
-{
-    "nombre":"Natali Suarez",
-    "usuario":"nati",
-    "password":"8765",
-    "role":"USER",
+  },
+  {
+    "nombre": "Natali Suarez",
+    "usuario": "nati",
+    "password": "8765",
+    "role": "USER",
     "isAuthenticated": true,
     "loginError": false,
     "img": "https://github.com/natalicmk.png"
-},
-{
-    "nombre":"invitado",
-    "usuario":"invitado",
-    "password":"",
-    "role":"INVITADO",
+  },
+  {
+    "nombre": "invitado",
+    "usuario": "invitado",
+    "password": "",
+    "role": "INVITADO",
     "isAuthenticated": true,
     "loginError": false,
     "img": ""
-}]
+  }
+]
